// samples/openapi3/client/petstore/typescript/builds/explode-query/models/Client.ts
/**
 * OpenAPI Petstore
 * This spec is mainly for testing Petstore server and contains fake endpoints, models. Please do not use this for any other purpose. Special characters: \" \\
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { AttributeTypeMapEntry } from '../models/ModelTypes';
import { HttpFile } from '../http/http';

export class Client {
    'client'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<AttributeTypeMapEntry> = [
        {
            "name": "client",
            "baseName": "client",
            "type": "string",
            "format": "",
            "required": false
        }    ];

    static getAttributeTypeMap() {
        return Client.attributeTypeMap;
    }

    public constructor() {
    }
}